requirements.txt
streamlit
opencv-python-headless
python-dotenv
python-docx
PyMuPDF
groq